Move non-view components out of views/ and group transit by submodule
views/ held ~17 components that no route renders: the directions trip and
waypoint UI, the timeline rows and charts (under a components/ directory nested
inside views/), the settings navigation and pane, the billing section and the
dialog host. Each moves to its feature module.

directions/timeline/ is renamed to directions/trip/ so it stops colliding with
the timeline location-history module, and DepartureBoard moves to transit.
The six flat ServiceAlert* files, RouteBullet and the departures components are
grouped under transit/{alerts,bullets,departures}/.
