+++ web/src/App.vue
@@ -39,7 +39,7 @@ import MobileNav from '@/components/navigation/MobileNavigation.vue'
 import FeedbackDialog from '@/components/feedback/FeedbackDialog.vue'
 import { useFeedback } from '@/composables/useFeedback'
 import { useShakeGesture } from '@/composables/useShakeGesture'
-import DialogView from '@/views/DialogView.vue'
+import DialogView from '@/components/shell/DialogHost.vue'
 import HotkeysMenu from '@/components/HotkeysMenu.vue'
 import ImpersonationBanner from '@/components/ImpersonationBanner.vue'
 import OnboardingDialog from '@/components/onboarding/OnboardingDialog.vue'
