+++ web/src/components/directions/TripItem.vue
