+++ web/src/components/directions/trip/SegmentDetails.vue
