+++ web/src/components/palette/Palette.vue
@@ -32,7 +32,7 @@ import {
   SettingsIcon,
 } from 'lucide-vue-next'
 import { ItemIcon } from '@/components/ui/item-icon'
-import RouteBullet from '@/components/transit/RouteBullet.vue'
+import RouteBullet from '@/components/transit/bullets/RouteBullet.vue'
 import { PlaceCard } from '@/components/place/card'
 import { makePlaceDisplay } from '@/lib/place-display'
 import { Badge } from '@/components/ui/badge'
