+++ web/src/components/timeline/DailyDistanceChart.vue
