+++ web/src/components/transit/departures/RealtimeIndicator.vue
