+++ web/src/components/transit/departures/StationTransfers.vue
@@ -15,12 +15,12 @@
  */
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
-import RouteBullet from '@/components/transit/RouteBullet.vue'
+import RouteBullet from '@/components/transit/bullets/RouteBullet.vue'
 import { bulletFor } from '@/services/layers/features/portolan/portolan-bullets'
 import { getRouteBulletLabel } from '@/lib/transit'
 import type { StationLine } from '@/composables/usePlaceTransitLines'
 import type { Dayjs } from 'dayjs'
-import RealtimeIndicator from '@/components/transit/RealtimeIndicator.vue'
+import RealtimeIndicator from '@/components/transit/departures/RealtimeIndicator.vue'
 import { ChevronRightIcon } from 'lucide-vue-next'
 import { formatCountdown, type RouteGroup } from '@/lib/transit-departures'
