+++ web/src/views/directions/Directions.vue
@@ -20,10 +20,10 @@ import {
 } from 'lucide-vue-next'
 import { useDirectionsStore, RIDESHARE_ENABLED } from '@/stores/directions.store'
 import { storeToRefs } from 'pinia'
-import WaypointInput from './WaypointInput.vue'
-import TripsList from './TripsList.vue'
-import RoutingPreferences from './RoutingPreferences.vue'
-import DirectionsLoading from './DirectionsLoading.vue'
+import WaypointInput from '@/components/directions/WaypointInput.vue'
+import TripsList from '@/components/directions/TripsList.vue'
+import RoutingPreferences from '@/components/directions/preferences/RoutingPreferences.vue'
+import DirectionsLoading from '@/components/directions/DirectionsLoading.vue'
 import { useElementSize } from '@vueuse/core'
 import { useSheetPeek } from '@/composables/useSheetPeek'
 import { Waypoint } from '@/types/map.types'
