+++ web/src/views/directions/TripDetail.vue
@@ -53,12 +53,12 @@ import { useThemeStore } from '@/stores/theme.store'
 import { ItemIcon } from '@/components/ui/item-icon'
 import { PlaceCard } from '@/components/place/card'
 import { waypointToDisplay, type PlaceDisplay } from '@/lib/place-display'
-import SegmentDetails from '@/components/directions/timeline/SegmentDetails.vue'
-import RealtimeIndicator from '@/components/transit/RealtimeIndicator.vue'
-import RouteBullet from '@/components/transit/RouteBullet.vue'
-import DepartureBoard from '@/components/directions/timeline/DepartureBoard.vue'
-import ServiceAlertRow from '@/components/transit/ServiceAlertRow.vue'
-import ServiceAlertCard from '@/components/transit/ServiceAlertCard.vue'
+import SegmentDetails from '@/components/directions/trip/SegmentDetails.vue'
+import RealtimeIndicator from '@/components/transit/departures/RealtimeIndicator.vue'
+import RouteBullet from '@/components/transit/bullets/RouteBullet.vue'
+import DepartureBoard from '@/components/transit/departures/DepartureBoard.vue'
+import ServiceAlertRow from '@/components/transit/alerts/ServiceAlertRow.vue'
+import ServiceAlertCard from '@/components/transit/alerts/ServiceAlertCard.vue'
 import { useTransitAlertsStore } from '@/stores/transit-alerts.store'
 import { splitFeedId, isInEffect, sortByRelevance } from '@/lib/transit-alerts'
 import type { ServiceAlert } from '@/types/transit.types'
