+++ web/src/views/settings/Settings.vue
@@ -4,8 +4,8 @@ import { useRouter } from 'vue-router'
 import { useResponsive } from '@/lib/utils'
 import { useSettingsScrollTarget } from '@/composables/useSettingsScrollTarget'
 
-import Navigation from './Navigation.vue'
-import SettingsPage from './SettingsPage.vue'
+import Navigation from '@/components/settings/SettingsNavigation.vue'
+import SettingsPage from '@/components/settings/SettingsPane.vue'
 
 const router = useRouter()
 const { isMobileScreen } = useResponsive()
