+++ web/src/views/settings/pages/Account.vue
@@ -14,7 +14,7 @@ import { Check, Pencil } from 'lucide-vue-next'
 import Passkeys from '@/components/auth/Passkeys.vue'
 import Sessions from '@/components/auth/Sessions.vue'
 import IdentitySettings from '@/components/identity/IdentitySettings.vue'
-import Billing from '@/views/settings/pages/Billing.vue'
+import Billing from '@/components/subscription/BillingSection.vue'
 
 const authService = useAuthService()
 const authStore = useAuthStore()
