+++ web/src/views/timeline/Timeline.vue
@@ -13,11 +13,11 @@ import { useIntegrationsStore } from '@/stores/integrations.store'
 import { useTimelineStore } from '@/stores/timeline.store'
 import { useMapService } from '@/services/map.service'
 import type { LocationHistoryStop, LocationHistoryEntry } from '@server/types/location-history.types'
-import TimelineStopRow from './components/TimelineStopRow.vue'
-import TimelineSegmentRow from './components/TimelineSegmentRow.vue'
-import DailyDistanceChart from './components/DailyDistanceChart.vue'
-import TimelineDatePicker from './components/TimelineDatePicker.vue'
-import TimelineNoIntegration from './TimelineNoIntegration.vue'
+import TimelineStopRow from '@/components/timeline/TimelineStopRow.vue'
+import TimelineSegmentRow from '@/components/timeline/TimelineSegmentRow.vue'
+import DailyDistanceChart from '@/components/timeline/DailyDistanceChart.vue'
+import TimelineDatePicker from '@/components/timeline/TimelineDatePicker.vue'
+import TimelineNoIntegration from '@/components/timeline/TimelineNoIntegration.vue'
 import { useAuthService } from '@/services/auth.service'
 import { PermissionId } from '@/types/auth.types'
 import UpgradeBanner from '@/components/subscription/UpgradeBanner.vue'
